+++ simulation/traffic_simulator/include/traffic_simulator/entity/vehicle_entity.hpp
@@ -75,7 +75,9 @@ public:
 
   void requestAcquirePosition(const geometry_msgs::msg::Pose & map_pose) override;
 
-  void requestLaneChange(const std::int64_t to_lanelet_id);
+  void requestLaneChange(const std::int64_t to_lanelet_id) override;
+
+  void requestLaneChange(const traffic_simulator::lane_change::Parameter & parameter) override;
 
   void cancelRequest() override;