+++ simulation/traffic_simulator/src/api/api.cpp
@@ -460,9 +460,31 @@ void API::requestLaneChange(const std::string & name, const std::int64_t & lanel
 }
 
 void API::requestLaneChange(
-  const std::string & name, const traffic_simulator::entity::Direction & direction)
+  const std::string & name, const traffic_simulator::lane_change::Direction & direction)
 {
   entity_manager_ptr_->requestLaneChange(name, direction);
+}
+
+void API::requestLaneChange(
+  const std::string & name, const traffic_simulator::lane_change::Parameter & parameter)
+{
+  entity_manager_ptr_->requestLaneChange(name, parameter);
+}
+
+void API::requestLaneChange(
+  const std::string & name, const traffic_simulator::lane_change::RelativeTarget & target,
+  const traffic_simulator::lane_change::Trajectory trajectory,
+  const lane_change::Constraint & constraint)
+{
+  entity_manager_ptr_->requestLaneChange(name, target, trajectory, constraint);
+}
+
+void API::requestLaneChange(
+  const std::string & name, const traffic_simulator::lane_change::AbsoluteTarget & target,
+  const traffic_simulator::lane_change::Trajectory trajectory,
+  const lane_change::Constraint & constraint)
+{
+  entity_manager_ptr_->requestLaneChange(name, target, trajectory, constraint);
 }
 
 }  // namespace traffic_simulator