+++ simulation/traffic_simulator/src/entity/entity_base.cpp
@@ -82,6 +82,53 @@ void EntityBase::requestSpeedChange(
   }
 }
 
+void EntityBase::requestLaneChange(
+  const traffic_simulator::lane_change::AbsoluteTarget & target,
+  const traffic_simulator::lane_change::Trajectory trajectory,
+  const lane_change::Constraint & constraint)
+{
+  auto param = traffic_simulator::lane_change::Parameter(target, trajectory, constraint);
+  requestLaneChange(param);
+}
+
+void EntityBase::requestLaneChange(
+  const traffic_simulator::lane_change::RelativeTarget & target,
+  const traffic_simulator::lane_change::Trajectory trajectory,
+  const lane_change::Constraint & constraint)
+{
+  std::int64_t reference_lanelet_id = 0;
+  if (target.entity_name == name) {
+    if (!getStatus().lanelet_pose_valid) {
+      THROW_SEMANTIC_ERROR(
+        "Target entity does not assigned to lanelet. Please check Target entity name : ",
+        target.entity_name, " exists on lane.");
+    }
+    reference_lanelet_id = getStatus().lanelet_pose.lanelet_id;
+  } else {
+    if (other_status_.find(target.entity_name) == other_status_.end()) {
+      THROW_SEMANTIC_ERROR(
+        "Target entity : ", target.entity_name, " does not exist. Please check ",
+        target.entity_name, " exists.");
+    }
+    if (!other_status_.at(target.entity_name).lanelet_pose_valid) {
+      THROW_SEMANTIC_ERROR(
+        "Target entity does not assigned to lanelet. Please check Target entity name : ",
+        target.entity_name, " exists on lane.");
+    }
+    reference_lanelet_id = other_status_.at(target.entity_name).lanelet_pose.lanelet_id;
+  }
+  const auto lane_change_target_id = hdmap_utils_ptr_->getLaneChangeableLaneletId(
+    reference_lanelet_id, target.direction, target.shift);
+  if (lane_change_target_id) {
+    requestLaneChange(
+      traffic_simulator::lane_change::AbsoluteTarget(lane_change_target_id.get(), target.offset),
+      trajectory, constraint);
+  } else {
+    THROW_SEMANTIC_ERROR(
+      "Failed to calculate absolute target lane. Please check the target lane exists.");
+  }
+}
+
 const autoware_vehicle_msgs::msg::VehicleCommand EntityBase::getVehicleCommand()
 {
   THROW_SIMULATION_ERROR("get vehicle command does not support in ", type, " entity type");