+++ docs/ReleaseNotes.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Difference between the latest release and master
+- Start supporting linear trajectory shape while changing lane. ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/661))
 - Fix context panel to display simulation time. (Contribution by [Utaro-M](https://github.com/Utaro-M)).
 - Support new vehicle_model_type `DELAY_STEER_VEL` ([pull request](https://github.com/tier4/scenario_simulator_v2/pull/660)).